+++ app/Http/Controllers/Admin/Gostaresh/EmployeeProfileController.php
@@ -6,6 +6,7 @@ use App\Exports\Gostaresh\EmployeeProfile\ListExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Gostaresh\EmployeeProfile\EmployeeProfileRequest;
 use App\Models\Index\EmployeeProfile;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
@@ -16,10 +17,12 @@ class EmployeeProfileController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-EmployeeProfile");
+
         $query = EmployeeProfile::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = EmployeeProfile::$filterColumnsCheckBoxes;
@@ -65,14 +68,16 @@ class EmployeeProfileController extends Controller
         return view('admin.gostaresh.employee-profile.list.pdf', compact('employeeProfiles'));
     }
     // ****************** End Export ******************
-    
+
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-EmployeeProfile");
+
         return view('admin.gostaresh.employee-profile.create.create');
     }
 
@@ -80,10 +85,11 @@ class EmployeeProfileController extends Controller
      * Store a newly created resource in storage.
      *
      * @param EmployeeProfileRequest $request
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function store(EmployeeProfileRequest $request)
     {
+        $this->authorize("create-any-EmployeeProfile");
         EmployeeProfile::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return redirect()->back()->with('success', __('titles.success_store'));
     }
@@ -103,11 +109,14 @@ class EmployeeProfileController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param EmployeeProfile $employeeProfile
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function edit(EmployeeProfile $employeeProfile)
     {
-        return view('admin.gostaresh.employee-profile.edit.edit', compact('employeeProfile'));
+        $this->authorize("edit-any-EmployeeProfile");
+
+        return view(
+            'admin.gostaresh.employee-profile.edit.edit', compact('employeeProfile'));
     }
 
     /**
@@ -115,11 +124,13 @@ class EmployeeProfileController extends Controller
      *
      * @param EmployeeProfileRequest $request
      * @param EmployeeProfile $employeeProfile
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function update(EmployeeProfileRequest $request, EmployeeProfile $employeeProfile)
     {
+        $this->authorize("edit-any-EmployeeProfile");
         $employeeProfile->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
@@ -127,10 +138,11 @@ class EmployeeProfileController extends Controller
      * Remove the specified resource from storage.
      *
      * @param EmployeeProfile $employeeProfile
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function destroy(EmployeeProfile $employeeProfile)
     {
+        $this->authorize("delete-any-EmployeeProfile");
         $employeeProfile->delete();
         return back()->with('success', __('titles.success_delete'));
     }
