+++ app/Http/Controllers/Admin/Gostaresh/EmploymentOfProvincialController.php
@@ -4,10 +4,10 @@ namespace App\Http\Controllers\Admin\Gostaresh;
 
 use App\Exports\Gostaresh\EmploymentOfProvincial\ListExport;
 use App\Http\Controllers\Controller;
-use App\Models\Index\EmploymentOfProvincial;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Gostaresh\EmploymentOfProvincial\EmploymentOfProvincialRequest;
+use App\Models\Index\EmploymentOfProvincial;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -17,10 +17,12 @@ class EmploymentOfProvincialController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-EmploymentOfProvincial");
+
         $query = EmploymentOfProvincial::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = EmploymentOfProvincial::$filterColumnsCheckBoxes;
@@ -64,21 +66,24 @@ class EmploymentOfProvincialController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-EmploymentOfProvincial");
+
         return view('admin.gostaresh.employment-of-provincial.create.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  EmploymentOfProvincialRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param EmploymentOfProvincialRequest $request
+     * @return Response
      */
     public function store(EmploymentOfProvincialRequest $request)
     {
+        $this->authorize("create-any-EmploymentOfProvincial");
         EmploymentOfProvincial::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return back()->with('success', __('titles.success_store'));
     }
@@ -86,8 +91,8 @@ class EmploymentOfProvincialController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function show(EmploymentOfProvincial $employmentOfProvincial)
     {
@@ -97,35 +102,41 @@ class EmploymentOfProvincialController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function edit(EmploymentOfProvincial $employmentOfProvincial)
     {
-        return view('admin.gostaresh.employment-of-provincial.edit.edit', compact('employmentOfProvincial'));
+        $this->authorize("edit-any-EmploymentOfProvincial");
+
+        return view(
+            'admin.gostaresh.employment-of-provincial.edit.edit', compact('employmentOfProvincial'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  EmploymentOfProvincialRequest  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param EmploymentOfProvincialRequest $request
+     * @param int $id
+     * @return Response
      */
     public function update(EmploymentOfProvincialRequest $request, EmploymentOfProvincial $employmentOfProvincial)
     {
+        $this->authorize("edit-any-EmploymentOfProvincial");
         $employmentOfProvincial->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function destroy(EmploymentOfProvincial $employmentOfProvincial)
     {
+        $this->authorize("delete-any-EmploymentOfProvincial");
         $employmentOfProvincial->delete();
         return back()->with('success', __('titles.success_delete'));
     }
