+++ app/Http/Controllers/Admin/Gostaresh/GeographicalLocationOfUnitController.php
@@ -4,13 +4,10 @@ namespace App\Http\Controllers\Admin\Gostaresh;
 
 use App\Exports\Gostaresh\GeographicalLocationOfUnit\ListExport;
 use App\Http\Controllers\Controller;
-use App\Models\Index\GeographicalLocationOfUnit;
-use Facades\Verta;
-use Illuminate\Contracts\View\View;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\View as FacadesView;
 use App\Http\Requests\Gostaresh\GeographicalLocationOfUnit\GeographicalLocationOfUnitRequest;
+use App\Models\Index\GeographicalLocationOfUnit;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -20,10 +17,12 @@ class GeographicalLocationOfUnitController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-GeographicalLocationOfUnit");
+
         $query = GeographicalLocationOfUnit::whereRequestsQuery();
         $geographicalLocationOfUnits = $query->orderBy('id', 'desc')->paginate(20);
         return view('admin.gostaresh.geographical-location-of-unit.list.list', compact('geographicalLocationOfUnits'));
@@ -57,21 +56,24 @@ class GeographicalLocationOfUnitController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-GeographicalLocationOfUnit");
+
         return view('admin.gostaresh.geographical-location-of-unit.create.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  GeographicalLocationOfUnitRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param GeographicalLocationOfUnitRequest $request
+     * @return Response
      */
     public function store(GeographicalLocationOfUnitRequest $request)
     {
+        $this->authorize("create-any-GeographicalLocationOfUnit");
         GeographicalLocationOfUnit::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return redirect()->back()->with('success', __('titles.success_store'));
     }
@@ -79,8 +81,8 @@ class GeographicalLocationOfUnitController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function show(GeographicalLocationOfUnit $geographicalLocationOfUnit)
     {
@@ -90,35 +92,41 @@ class GeographicalLocationOfUnitController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function edit(GeographicalLocationOfUnit $geographicalLocationOfUnit)
     {
-        return view('admin.gostaresh.geographical-location-of-unit.edit.edit', compact('geographicalLocationOfUnit'));
+        $this->authorize("edit-any-GeographicalLocationOfUnit");
+
+        return view(
+            'admin.gostaresh.geographical-location-of-unit.edit.edit', compact('geographicalLocationOfUnit'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  GeographicalLocationOfUnitRequest  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param GeographicalLocationOfUnitRequest $request
+     * @param int $id
+     * @return Response
      */
     public function update(GeographicalLocationOfUnitRequest $request, GeographicalLocationOfUnit $geographicalLocationOfUnit)
     {
+        $this->authorize("edit-any-GeographicalLocationOfUnit");
         $geographicalLocationOfUnit->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function destroy(GeographicalLocationOfUnit $geographicalLocationOfUnit)
     {
+        $this->authorize("delete-any-GeographicalLocationOfUnit");
         $geographicalLocationOfUnit->delete();
         return back()->with('success', __('titles.success_delete'));
     }
