+++ app/Http/Controllers/Admin/Gostaresh/GraduatesOfHigherEducationController.php
@@ -6,7 +6,7 @@ use App\Exports\Gostaresh\GraduatesOfHigherEducation\ListExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Gostaresh\GraduatesOfHigherEducation\GraduatesOfHigherEducationRequest;
 use App\Models\Index\GraduatesOfHigherEducationCenters;
-use App\Models\Index\StatusAnalysisOfTheNumberOfCurricula;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
@@ -25,6 +25,8 @@ class GraduatesOfHigherEducationController extends Controller
      */
     public function index()
     {
+        $this->authorize("view-any-GraduatesOfHigherEducationCenters");
+
         $query = GraduatesOfHigherEducationCenters::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = GraduatesOfHigherEducationCenters::$filterColumnsCheckBoxes;
@@ -36,7 +38,7 @@ class GraduatesOfHigherEducationController extends Controller
         $graduatesOfHigherEducationCenters = $query->orderBy('id', 'desc')->paginate(20);
 
         return view('admin.gostaresh.graduates-of-higher-education.list.list', compact('graduatesOfHigherEducationCenters',
-            'yearSelectedList','filterColumnsCheckBoxes'));
+            'yearSelectedList', 'filterColumnsCheckBoxes'));
     }
 
     private function yearSelectedList($query)
@@ -77,6 +79,8 @@ class GraduatesOfHigherEducationController extends Controller
      */
     public function create()
     {
+        $this->authorize("create-any-GraduatesOfHigherEducationCenters");
+
         return view('admin.gostaresh.graduates-of-higher-education.create.create');
     }
 
@@ -85,9 +89,11 @@ class GraduatesOfHigherEducationController extends Controller
      *
      * @param GraduatesOfHigherEducationRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(GraduatesOfHigherEducationRequest $request)
     {
+        $this->authorize("create-any-GraduatesOfHigherEducationCenters");
         GraduatesOfHigherEducationCenters::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return redirect()->back()->with('success', __('titles.success_store'));
     }
@@ -111,7 +117,10 @@ class GraduatesOfHigherEducationController extends Controller
      */
     public function edit(GraduatesOfHigherEducationCenters $graduatesOfHigherEducation)
     {
-        return view('admin.gostaresh.graduates-of-higher-education.edit.edit', compact('graduatesOfHigherEducation'));
+        $this->authorize("edit-any-GraduatesOfHigherEducationCenters");
+
+        return view(
+            'admin.gostaresh.graduates-of-higher-education.edit.edit', compact('graduatesOfHigherEducation'));
     }
 
     /**
@@ -123,7 +132,9 @@ class GraduatesOfHigherEducationController extends Controller
      */
     public function update(GraduatesOfHigherEducationRequest $request, GraduatesOfHigherEducationCenters $graduatesOfHigherEducation)
     {
+        $this->authorize("edit-any-GraduatesOfHigherEducationCenters");
         $graduatesOfHigherEducation->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
@@ -135,6 +146,7 @@ class GraduatesOfHigherEducationController extends Controller
      */
     public function destroy(GraduatesOfHigherEducationCenters $graduatesOfHigherEducation)
     {
+        $this->authorize("delete-any-GraduatesOfHigherEducationCenters");
         $graduatesOfHigherEducation->delete();
         return back()->with('success', __('titles.success_delete'));
     }
