+++ app/Http/Controllers/Admin/Gostaresh/MultipleDeprivationIndexOfCityController.php
@@ -4,10 +4,10 @@ namespace App\Http\Controllers\Admin\Gostaresh;
 
 use App\Exports\Gostaresh\MultipleDeprivationIndexOfCity\ListExport;
 use App\Http\Controllers\Controller;
-use App\Models\Index\MultipleDeprivationIndexOfCity;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Gostaresh\MultipleDeprivationIndexOfCity\MultipleDeprivationIndexOfCityRequest;
+use App\Models\Index\MultipleDeprivationIndexOfCity;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -17,10 +17,12 @@ class MultipleDeprivationIndexOfCityController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-MultipleDeprivationIndexOfCity");
+
         $query = MultipleDeprivationIndexOfCity::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = MultipleDeprivationIndexOfCity::$filterColumnsCheckBoxes;
@@ -64,30 +66,35 @@ class MultipleDeprivationIndexOfCityController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-MultipleDeprivationIndexOfCity");
+
         return view('admin.gostaresh.multiple-deprivation-index-of-city.create.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  MultipleDeprivationIndexOfCityRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param MultipleDeprivationIndexOfCityRequest $request
+     * @return Response
      */
     public function store(MultipleDeprivationIndexOfCityRequest $request)
     {
+        $this->authorize("create-any-MultipleDeprivationIndexOfCity");
+
         MultipleDeprivationIndexOfCity::create(array_merge(['user_id' => Auth::id()], $request->validated()));
+
         return back()->with('success', __('titles.success_store'));
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function show(MultipleDeprivationIndexOfCity $multipleDeprivationIndexOfCity)
     {
@@ -97,35 +104,41 @@ class MultipleDeprivationIndexOfCityController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function edit(MultipleDeprivationIndexOfCity $multipleDeprivationIndexOfCity)
     {
-        return view('admin.gostaresh.multiple-deprivation-index-of-city.edit.edit', compact('multipleDeprivationIndexOfCity'));
+        $this->authorize("edit-any-MultipleDeprivationIndexOfCity");
+
+        return view(
+            'admin.gostaresh.multiple-deprivation-index-of-city.edit.edit', compact('multipleDeprivationIndexOfCity'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  MultipleDeprivationIndexOfCityRequest  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param MultipleDeprivationIndexOfCityRequest $request
+     * @param int $id
+     * @return Response
      */
     public function update(MultipleDeprivationIndexOfCityRequest $request, MultipleDeprivationIndexOfCity $multipleDeprivationIndexOfCity)
     {
+        $this->authorize("edit-any-MultipleDeprivationIndexOfCity");
         $multipleDeprivationIndexOfCity->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function destroy(MultipleDeprivationIndexOfCity $multipleDeprivationIndexOfCity)
     {
+        $this->authorize("delete-any-MultipleDeprivationIndexOfCity");
         $multipleDeprivationIndexOfCity->delete();
         return back()->with('success', __('titles.success_delete'));
     }
